requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.3.2
langchain
langchain-community
langchain-google-genai
langchain-chroma
chromadb>=1.0.0
tiktoken==0.5.2
rank-bm25==0.2.2
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
scikit-learn>=1.5.0
duckdb==0.10.0
sentence-transformers>=2.2.2
python-docx>=0.8.11
langchain-core>=0.1.0